+++ zerver/tests/test_presence.py
@@ -9,64 +9,15 @@ from zerver.lib.actions import do_deactivate_user
 from zerver.lib.presence import get_status_dict_by_realm
 from zerver.lib.statistics import seconds_usage_between
 from zerver.lib.test_classes import ZulipTestCase
-from zerver.lib.test_helpers import make_client, queries_captured, reset_emails_in_zulip_realm
+from zerver.lib.test_helpers import make_client, reset_emails_in_zulip_realm
 from zerver.lib.timestamp import datetime_to_timestamp
 from zerver.models import (
-    Client,
     PushDeviceToken,
     UserActivity,
     UserActivityInterval,
     UserPresence,
     UserProfile,
-    flush_per_request_caches,
 )
-
-
-class ActivityTest(ZulipTestCase):
-    @mock.patch("stripe.Customer.list", return_value=[])
-    def test_activity(self, unused_mock: mock.Mock) -> None:
-        self.login("hamlet")
-        client, _ = Client.objects.get_or_create(name="website")
-        query = "/json/messages/flags"
-        last_visit = timezone_now()
-        count = 150
-        for activity_user_profile in UserProfile.objects.all():
-            UserActivity.objects.get_or_create(
-                user_profile=activity_user_profile,
-                client=client,
-                query=query,
-                count=count,
-                last_visit=last_visit,
-            )
-
-        # Fails when not staff
-        result = self.client_get("/activity")
-        self.assertEqual(result.status_code, 302)
-
-        user_profile = self.example_user("hamlet")
-        user_profile.is_staff = True
-        user_profile.save(update_fields=["is_staff"])
-
-        flush_per_request_caches()
-        with queries_captured() as queries:
-            result = self.client_get("/activity")
-            self.assertEqual(result.status_code, 200)
-
-        self.assert_length(queries, 19)
-
-        flush_per_request_caches()
-        with queries_captured() as queries:
-            result = self.client_get("/realm_activity/zulip/")
-            self.assertEqual(result.status_code, 200)
-
-        self.assert_length(queries, 8)
-
-        flush_per_request_caches()
-        with queries_captured() as queries:
-            result = self.client_get("/user_activity/iago@zulip.com/")
-            self.assertEqual(result.status_code, 200)
-
-        self.assert_length(queries, 4)
 
 
 class TestClientModel(ZulipTestCase):
